Update clean_text preprocessor and unit tests with standardized stopword behavior
With the latest Pew Analytics, empty lists passed to a TextCleaner's stopwords parameter will result in no stopwords being used. This behavior should replicate in Django Learning, especially since "english" is a built-in, and optional, stopword set; opting to not pass "english" to stopword_sets in a pipeline should not result in English stopwords being used anyway.